requirements.txt
Django==2.1.5
djangorestframework==3.9.1
djangorestframework-simplejwt==3.3
mysqlclient==1.4.1
PyJWT==1.7.1
pytz==2018.9
